Refactored routes, requests now don't need to pass the user id anymore

+++ src/main/kotlin/com/thewyp/routes/CommentRoutes.kt
@@ -3,6 +3,7 @@ package com.thewyp.routes
 import com.thewyp.data.requests.CreateCommentRequest
 import com.thewyp.data.requests.DeleteCommentRequest
 import com.thewyp.data.responses.BasicApiResponse
+import com.thewyp.plugins.userId
 import com.thewyp.service.CommentService
 import com.thewyp.service.LikeService
 import com.thewyp.service.UserService
@@ -16,8 +17,7 @@ import io.ktor.response.*
 import io.ktor.routing.*
 
 fun Route.createComment(
-    commentService: CommentService,
-    userService: UserService
+    commentService: CommentService
 ) {
     authenticate {
         post("/api/comment/create") {
@@ -25,37 +25,33 @@ fun Route.createComment(
                 call.respond(HttpStatusCode.BadRequest)
                 return@post
             }
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ) {
-                when(commentService.createComment(request)) {
-                    is CommentService.ValidationEvent.ErrorFieldEmpty -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = false,
-                                message = ApiResponseMessages.FIELDS_BLANK
-                            )
+            println("createComment:userId=${call.userId}")
+            when(commentService.createComment(request, call.userId)) {
+                is CommentService.ValidationEvent.ErrorFieldEmpty -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = false,
+                            message = ApiResponseMessages.FIELDS_BLANK
                         )
-                    }
-                    is CommentService.ValidationEvent.ErrorCommentTooLong -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = false,
-                                message = ApiResponseMessages.COMMENT_TOO_LONG
-                            )
+                    )
+                }
+                is CommentService.ValidationEvent.ErrorCommentTooLong -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = false,
+                            message = ApiResponseMessages.COMMENT_TOO_LONG
                         )
-                    }
-                    is CommentService.ValidationEvent.Success -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = true,
-                            )
+                    )
+                }
+                is CommentService.ValidationEvent.Success -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = true,
                         )
-                    }
+                    )
                 }
             }
         }
@@ -79,7 +75,6 @@ fun Route.getCommentsForPost(
 
 fun Route.deleteComment(
     commentService: CommentService,
-    userService: UserService,
     likeService: LikeService
 ) {
     authenticate {
@@ -88,17 +83,18 @@ fun Route.deleteComment(
                 call.respond(HttpStatusCode.BadRequest)
                 return@delete
             }
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ) {
-                val deleted = commentService.deleteComment(request.commentId)
-                if(deleted) {
-                    likeService.deleteLikesForParent(request.commentId)
-                    call.respond(HttpStatusCode.OK, BasicApiResponse(successful = true))
-                } else {
-                    call.respond(HttpStatusCode.NotFound, BasicApiResponse(successful = false))
-                }
+            println("deleteComment:userId=${call.userId}")
+            val comment = commentService.getCommentById(request.commentId)
+            if(comment?.userId != call.userId) {
+                call.respond(HttpStatusCode.Unauthorized)
+                return@delete
+            }
+            val deleted = commentService.deleteComment(request.commentId)
+            if(deleted) {
+                likeService.deleteLikesForParent(request.commentId)
+                call.respond(HttpStatusCode.OK, BasicApiResponse(successful = true))
+            } else {
+                call.respond(HttpStatusCode.NotFound, BasicApiResponse(successful = false))
             }
         }
     }
